Remove unused rootless-cni-infra container files
